+++ tests/test_runner.py
@@ -1,90 +1,65 @@
-from dataclasses import dataclass, field
 from pathlib import Path
 from unittest.mock import Mock, patch
 
-import pytest
-
-from tests.conftest import ContextBuilderFunc
 from universal_test_runner.context import Context
 from universal_test_runner.runner import run, run_test_command
 
 
-def test_no_matches(capsys, build_context: ContextBuilderFunc):
-    assert run_test_command(build_context()) == 1
-    out, _ = capsys.readouterr()
-    assert "no testing method" in out
+@patch("subprocess.run")
+def test_run_command(subp_run: Mock):
+    subp_run.return_value = Mock(returncode=0)
 
+    assert run_test_command(["a", "-b", "--c"]) == 0
 
-@dataclass
-class RunnerTestCase:
-    files: list[str]
-    expected_command: str
-    args: list[str] = field(default_factory=list)
-
-    def __repr__(self) -> str:
-        return f"files={self.files} & args={self.args} -> `{self.expected_command}`"
+    subp_run.assert_called_once_with(["a", "-b", "--c"])
 
 
 @patch("subprocess.run")
-class TestRunner:
-    def test_command_not_found(
-        self, subp_run: Mock, capsys, build_context: ContextBuilderFunc
-    ):
-        subp_run.side_effect = FileNotFoundError
+def test_no_matches(subp_run: Mock, capsys):
+    assert run_test_command([]) == 1
 
-        assert run_test_command(build_context([".pytest_cache"])) == 1
+    out, _ = capsys.readouterr()
+    assert "no testing method found" in out
 
-        out, _ = capsys.readouterr()
-        assert "command not found:" in out
-
-    @pytest.mark.parametrize(
-        "test_case",
-        [
-            RunnerTestCase([".pytest_cache"], "pytest"),
-            RunnerTestCase(["tests.py"], "python tests.py"),
-            RunnerTestCase(["go.mod"], "go test ./..."),
-            RunnerTestCase(
-                ["go.mod"], args=["parser"], expected_command="go test parser"
-            ),
-            RunnerTestCase(["parser_test.go"], "go test"),
-            RunnerTestCase(
-                ["parser_test.go"], args=["-v"], expected_command="go test -v"
-            ),
-            RunnerTestCase(["mix.exs"], "mix test"),
-            RunnerTestCase(["Cargo.toml"], "cargo test"),
-            RunnerTestCase(["project.clj"], "lein test"),
-        ],
-        ids=repr,
-    )
-    def test_matchers(
-        self,
-        subp_run: Mock,
-        test_case: RunnerTestCase,
-        build_context: ContextBuilderFunc,
-    ):
-        run_test_command(build_context(test_case.files, test_case.args))
-
-        subp_run.assert_called_once_with(test_case.expected_command.split())
+    subp_run.assert_not_called()
 
 
-@patch("sys.argv", new=["test-runner", "-a", "--b", "c"])
+@patch("subprocess.run")
+def test_command_not_found(subp_run: Mock, capsys):
+    subp_run.side_effect = FileNotFoundError
+
+    assert run_test_command(["pytest"]) == 1
+
+    out, _ = capsys.readouterr()
+    assert "command not found:" in out
+
+    subp_run.assert_called_once_with(["pytest"])
+
+
+@patch("sys.argv", new=["test-runner", "a", "-b", "--c"])
 @patch("sys.exit")
+@patch("universal_test_runner.runner.find_test_command")
 @patch("universal_test_runner.runner.run_test_command")
 @patch("os.getcwd")
 def test_run(
     mock_cwd: Mock,
-    mock_runner: Mock,
+    mock_test_runner: Mock,
+    mock_command_finder: Mock,
     mock_exit: Mock,
     tmp_path: Path,
     touch_files,
 ):
-    files = ["x.py", "y.py", "z.py"]
+    files = ["a.txt", "b.txt", "c.txt"]
     touch_files(files)
+
     mock_cwd.return_value = str(tmp_path)
+
+    mock_command_finder.return_value = ["my", "test", "command"]
 
     run()
 
-    mock_runner.assert_called_once_with(
-        Context(str(tmp_path), frozenset(files), ("-a", "--b", "c"))
+    mock_command_finder.assert_called_once_with(
+        Context(str(tmp_path), frozenset(files), ("a", "-b", "--c"))
     )
-    mock_exit.assert_called_once_with(mock_runner.return_value)
+    mock_test_runner.assert_called_once_with(["my", "test", "command"])
+    mock_exit.assert_called_once_with(mock_test_runner.return_value)