+++ universal_test_runner/runner.py
@@ -1,37 +1,29 @@
-import os
 import subprocess
 import sys
 
 from universal_test_runner.context import Context
-from universal_test_runner.matchers import ALL_MATCHERS
+from universal_test_runner.matchers import find_test_command
 
 
-def run_test_command(context: Context) -> int:
-    for matcher in ALL_MATCHERS:
-        if not matcher.matches(context):
-            continue
-
-        command = [*matcher.command, *context.args]
-
-        try:
-            return subprocess.run(command).returncode
-        except FileNotFoundError:
-            # e.g. if `pytest` is run, but not installed
-            # we capture the error so there's not a Python traceback shown
-            print(f"command not found: {command[0]}")
-            return 1
-    else:
+def run_test_command(command: list[str]) -> int:
+    if not command:
         print("no testing method found!")
+        return 1
+
+    try:
+        return subprocess.run(command).returncode
+    except FileNotFoundError:
+        # e.g. if `pytest` is run, but not installed
+        # we capture the error so there's not a Python traceback shown
+        print(f"command not found: {command[0]}")
         return 1
 
 
 # not a click handler, since this is just a passthrough for the underlying test runner
 def run():
-    current_dir = os.getcwd()
-
-    c = Context.build(current_dir, sys.argv[1:])
-
-    sys.exit(run_test_command(c))
+    context = Context.from_invocation()
+    command = find_test_command(context)
+    sys.exit(run_test_command(command))
 
 
 if __name__ == "__main__":